pyproject.toml
[tool.poetry]
name = "qwazii-service"
version = "0.1.0"
description = "Qwazii Service - A testing lecture sample project"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
gunicorn = "^23.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-django = "^4.5.2"
black = "^23.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
